data_pipeline/requirements.txt
requests
rasterio
numpy
shapely
geopandas
sentinelhub-py
sqlalchemy
psycopg2-binary
geoalchemy2
